Move get_destinations and get_additional_documents into GenericApplicationViewSerializer
