+++ applications/serializers/exhibition_clearance.py
@@ -1,9 +1,7 @@
 from rest_framework import serializers
 from rest_framework.fields import CharField
 
-from applications.models import ApplicationDocument
 from applications.models import ExhibitionClearanceApplication
-from applications.serializers.document import ApplicationDocumentSerializer
 from applications.serializers.generic_application import (
     GenericApplicationCreateSerializer,
     GenericApplicationViewSerializer,
@@ -13,7 +11,6 @@ from applications.serializers.good import GoodOnApplicationViewSerializer
 from cases.enums import CaseTypeEnum
 from lite_content.lite_api import strings
 from parties.serializers import (
-    EndUserSerializer,
     EndUserWithFlagsSerializer,
     UltimateEndUserWithFlagsSerializer,
     ThirdPartyWithFlagsSerializer,
@@ -43,17 +40,6 @@ class ExhibitionClearanceViewSerializer(GenericApplicationViewSerializer):
             "destinations",
             "additional_documents",
         )
-
-    def get_destinations(self, application):
-        if application.end_user:
-            serializer = EndUserSerializer(application.end_user)
-            return {"type": "end_user", "data": serializer.data}
-        else:
-            return {"type": "end_user", "data": ""}
-
-    def get_additional_documents(self, instance):
-        documents = ApplicationDocument.objects.filter(application=instance)
-        return ApplicationDocumentSerializer(documents, many=True).data
 
 
 class ExhibitionClearanceCreateSerializer(GenericApplicationCreateSerializer):
